moving structs about
