+++ pkg/platform/microservice/k8s_storage.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/dolittle-entropy/platform-api/pkg/platform"
 	"k8s.io/client-go/kubernetes"
 )
 
@@ -17,17 +18,17 @@ func NewK8sStorage(k8sClient *kubernetes.Clientset) *k8sStorage {
 	}
 }
 
-func (s *k8sStorage) Write(info HttpInputDolittle, data []byte) error {
+func (s *k8sStorage) Write(info platform.HttpInputDolittle, data []byte) error {
 	fmt.Printf("Write %s.json to file", info.MicroserviceID)
 	return errors.New("TODO")
 }
 
-func (s *k8sStorage) Read(info HttpInputDolittle) ([]byte, error) {
+func (s *k8sStorage) Read(info platform.HttpInputDolittle) ([]byte, error) {
 	fmt.Printf("%v", info)
 	return []byte(`{"TODO":"TODO"}`), nil
 }
 
-func (s *k8sStorage) GetAll(tenantID string, applicationID string) ([]HttpMicroserviceBase, error) {
-	services := make([]HttpMicroserviceBase, 0)
+func (s *k8sStorage) GetAll(tenantID string, applicationID string) ([]platform.HttpMicroserviceBase, error) {
+	services := make([]platform.HttpMicroserviceBase, 0)
 	return services, errors.New("TODO")
 }
